__BYPASS_CHECKS__

yt:git_commit:4d916ae0c6fe5069f8452118d3f4e4fc7a2a60f1
yt:last_git_svn_commit:71255c68f03f49ccf1c3f34722ca481fff96b916
yt:last_svn_revision:2770123

git-svn-id: svn+ssh://arcadia.yandex.ru/arc/trunk/arcadia/yt/19_2/yt@2792892 41d65440-b5be-11dd-afe3-b2e846d9b4f8

+++ core/yson/detail.h
@@ -67,7 +67,7 @@ template <>
 class TPositionInfo<true>
 {
 private:
-    int Offset;
+    i64 Offset;
     int Line;
     int Column;
 
@@ -103,7 +103,7 @@ template <>
 class TPositionInfo<false>
 {
 private:
-    int Offset;
+    i64 Offset;
 
 public:
     TPositionInfo()
